+++ client/src/conponents/common/header/navy/dropMenu/getCategoriesArray.js
@@ -86,3 +86,4 @@ parentId: "Gift Cards",
 description: "A category, of earrings",
 level: 0,
 date: "2020-02-10T14:54:54.951Z"},]
+
